fix tests and abi Tokenize overload
