+++ src/AngleSharp.Core.Tests/TestRuntime.cs
@@ -4,5 +4,6 @@ using System;
 
 public static class TestRuntime
 {
-    public static Boolean UsePrefetchedTextSource { get; set; } = Environment.GetEnvironmentVariable("prefetched") == "true";
+    public static Boolean UsePrefetchedTextSource { get; set; } =
+        Environment.GetEnvironmentVariable("prefetched") == "true";
 }
