+++ src/AngleSharp/Dom/Internal/Document.cs
@@ -34,8 +34,7 @@ namespace AngleSharp.Dom
         private readonly Window _view;
         private readonly IResourceLoader? _loader;
         private readonly Location _location;
-        private readonly IReadOnlyTextSource _source;
-        private readonly TextSource? _textSource;
+        private readonly TextSource _source;
         private readonly object _importedUrisLock = new object();
 
         private QuirksMode _quirksMode;
@@ -483,13 +482,7 @@ namespace AngleSharp.Dom
         #region ctor
 
         /// <inheritdoc />
-        public Document(IBrowsingContext context, TextSource source) : this(context, (IReadOnlyTextSource)source)
-        {
-            _textSource = source;
-        }
-
-        /// <inheritdoc />
-        public Document(IBrowsingContext context, IReadOnlyTextSource source)
+        public Document(IBrowsingContext context, TextSource source)
             : base(null, "#document", NodeType.Document)
         {
             Referrer = String.Empty;
@@ -515,12 +508,19 @@ namespace AngleSharp.Dom
             _statusCode = HttpStatusCode.OK;
         }
 
+        /// <inheritdoc />
+        public Document(IBrowsingContext context, IReadOnlyTextSource source)
+            : this(context, new TextSource(source))
+        {
+
+        }
+
         #endregion
 
         #region Properties
 
         /// <inheritdoc />
-        public TextSource Source => (_textSource ?? _source as TextSource)!;
+        public TextSource Source => (_source ?? _source as TextSource)!;
 
         /// <inheritdoc />
         public abstract IEntityProvider Entities
